bluetcl: carve the wiretypemap analysis out into a WireAnalysis module
Move the netlist wire-type analysis behind 'module wiretypemap' from
bluetcl.hs into its own WireAnalysis module, and use the aif_name field
directly rather than re-deriving it.

Provenance: re-cut of d24f09c59 + 11ed45f03 from claude/trs-fst-rebased.
